pyproject.toml
[project]
name = "backend-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "chromadb>=0.6.3",
    "fastapi>=0.115.12",
    "litellm>=1.69.2",
    "moviepy>=2.1.2",
    "openai>=1.71.0",
    "opencv-python>=4.11.0.86",
    "psutil>=7.0.0",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "tinydb>=4.8.2",
    "uvicorn>=0.34.0",
]
